requirements.txt
## How to set up a Python runtime environment for Vivarium.
#
## Create the "vivarium" python virtual environment and select it for your project
## directory:
#   cd ~/code/vivarium-core
#   pyenv local 3.8.3
#   pyenv virtualenv vivarium
#   pyenv local vivarium
#
## Install requirements
#   pip install -r requirements.txt
#

vivarium-core
tellurium 